+++ package/lib/src/beam_update_guard.dart
@@ -37,12 +37,13 @@ class BeamUpdateGuard {
   /// the one to which beaming has been requested.
   ///
   /// [context] is also injected to fetch data up the tree if necessary.
-  final bool Function(BeamLocation delegate, RouteInformation routeInformation, Object? data) check;
+  final bool Function(BeamLocation delegate, RouteInformation routeInformation,
+      Object? data) check;
 
   /// Arbitrary closure to execute when [check] fails.
-  final void Function(BeamerDelegate delegate, RouteInformation routeInformation, Object? data)
-      redirect;
-  
+  final void Function(BeamerDelegate delegate,
+      RouteInformation routeInformation, Object? data) redirect;
+
   /// Whether to [check] all the path blueprints defined in [pathPatterns]
   /// or [check] all the paths that **are not** in [pathPatterns].
   ///
@@ -55,13 +56,17 @@ class BeamUpdateGuard {
   /// contained within location's pathBlueprint.
   /// Else, the path (i.e. the pre-query substring) of the location's uri
   /// must be equal to the pathBlueprint.
-  bool _hasMatch(RouteInformation routeInformation) => patternsMatch(pathPatterns,routeInformation);
+  bool _hasMatch(RouteInformation routeInformation) =>
+      patternsMatch(pathPatterns, routeInformation);
 
   /// Whether or not the guard should [check] the [location].
   /// [currentLocation] is the current beam location. [routeInformation] and [data] have not yet applied to it.
   /// [routeInformation] is the route info of the new route
-  /// [data] is either data that was passed with the new route 
-  bool shouldGuard(BeamLocation currentLocation, RouteInformation routeInformation, Object? data) {
-    return guardNonMatching ? !_hasMatch(routeInformation) : _hasMatch(routeInformation);
+  /// [data] is either data that was passed with the new route
+  bool shouldGuard(BeamLocation currentLocation,
+      RouteInformation routeInformation, Object? data) {
+    return guardNonMatching
+        ? !_hasMatch(routeInformation)
+        : _hasMatch(routeInformation);
   }
 }